Refactor import statements in streamlit_fragments.py, view.py, and manager.py to improve module organization. Enhance error handling in filter_section function to provide user feedback on model generation failures.
